Fix participant not linked to parent contribution
 - prevent ContributionParticipations which are both authors and speakers to
   have their link to the contribution removed when the author is removed but
   the speaker remains.
 - fixes #1596

+++ indico/MaKaC/conference.py
@@ -8479,7 +8479,8 @@ class Contribution(CommonObjectBase, Locatable):
         if not self._authors.has_key( part.getId() ):
             return
         del self._authors[ part.getId() ]
-        part.delete()
+        if not self.isSpeaker(part):
+            part.delete()
 
     def addPrimaryAuthor(self, part, index=None):
         """
